+++ sei-ibc-go/modules/core/03-connection/module.go
@@ -13,11 +13,6 @@ func Name() string {
 	return types.SubModuleName
 }
 
-// GetTxCmd returns the root tx command for the IBC connections.
-func GetTxCmd() *cobra.Command {
-	return cli.NewTxCmd()
-}
-
 // GetQueryCmd returns the root query command for the IBC connections.
 func GetQueryCmd() *cobra.Command {
 	return cli.GetQueryCmd()